+++ src/Tokens/Text.php
@@ -3,6 +3,7 @@
     namespace Tholabs\I18nBuild\Tokens;
 
     class Text implements Tokenized {
+        use ChildlessTokenTrait;
 
         private string $body;