+++ src/Tokens/Variable.php
@@ -3,6 +3,7 @@
     namespace Tholabs\I18nBuild\Tokens;
 
     class Variable implements Tokenized {
+        use ChildlessTokenTrait;
 
         private string $variableName;